requirements.txt
python-dotenv
azure-identity
openai
azure-ai-projects==2.0.0b2